refactor: move wss to separate module
